Don't compile all LLVM tools, only the ones we are installing
Compiling this without ninja might be a bit more inefficient, but
should be a fair bit faster with ninja.

This reduces the llvm compile time by around 10 %.

+++ strip-llvm.sh
@@ -74,9 +74,11 @@ if [ -n "$EXEEXT" ]; then
 fi
 cd ..
 rm -rf share libexec
-cd include
-rm -rf clang clang-c lld llvm llvm-c
-cd ..
+if [ -d include ]; then
+    cd include
+    rm -rf clang clang-c lld llvm llvm-c
+    cd ..
+fi
 cd lib
 rm -rf lib*.a *.so* *.dylib* cmake
 cd ..
